+++ packages/web/src/types.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 export const orgMetadataSchema = z.object({
-    publicAccessEnabled: z.boolean().optional(),
+    anonymousAccessEnabled: z.boolean().optional(),
 })
 
 export type OrgMetadata = z.infer<typeof orgMetadataSchema>;